Consolidate SpectatorViewPlugin source into the SharedFiles directory
