--- src/SpectatorView.Native/SpectatorView.OpenCV/Desktop/dllmain.cpp
@@ -1,19 +1,0 @@
-#include "..\SharedFiles\pch.h"
-
-// dllmain.cpp : Defines the entry point for the DLL application.
-BOOL APIENTRY DllMain( HMODULE hModule,
-                       DWORD  ul_reason_for_call,
-                       LPVOID lpReserved
-                     )
-{
-    switch (ul_reason_for_call)
-    {
-    case DLL_PROCESS_ATTACH:
-    case DLL_THREAD_ATTACH:
-    case DLL_THREAD_DETACH:
-    case DLL_PROCESS_DETACH:
-        break;
-    }
-    return TRUE;
-}
-
