+++ src/SpectatorView.Native/SpectatorView.OpenCV/SharedFiles/SpectatorViewPlugin.h
