+++ src/SpectatorView.Native/SpectatorView.OpenCV/SharedFiles/dllmain.cpp
