+++ src/SpectatorView.Native/SpectatorView.OpenCV/UWP/SpectatorView.OpenCV.UWP.vcxproj
@@ -203,13 +203,13 @@
     <ClInclude Include="..\SharedFiles\ArUcoMarkerDetector.h" />
     <ClInclude Include="..\SharedFiles\Calibration.h" />
     <ClInclude Include="..\SharedFiles\pch.h" />
-    <ClInclude Include="SpectatorViewPlugin.h" />
+    <ClInclude Include="..\SharedFiles\SpectatorViewPlugin.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\SharedFiles\ArUcoMarkerDetector.cpp" />
     <ClCompile Include="..\SharedFiles\Calibration.cpp" />
-    <ClCompile Include="SpectatorViewPlugin.cpp" />
-    <ClCompile Include="dllmain.cpp" />
+    <ClCompile Include="..\SharedFiles\SpectatorViewPlugin.cpp" />
+    <ClCompile Include="..\SharedFiles\dllmain.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets" />
